+++ firmware/common2015/CMakeLists.txt
@@ -33,7 +33,7 @@ set(DRIVERS
     buzzer
     cc1201
     cc1201/cfg
-    decadriver
+    decawave
     ds2411
     fpga
     kicker-board
